+++ src/server/worldserver/TCSoap/TCSoap.cpp
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "OCSoap.h"
+#include "TCSoap.h"
 
 #define POOL_SIZE   5
